feat(ci): CI-running tests
This implements some lo2s tests that can and will actually run in the
CI.

While most of the perf hardware event based tests can not run, the
tracepoint based tests run, which should give us a pretty wide covering
of basic lo2s functionality at least.

In this step, I replace the fixtures with checks that make ctest skip
the test if the pre-requisites don't exist (For example if the event we
want to use does not exist per `perf`).

I acknowledge that we initially chosen fixtures for these pre-requisites
on purpose and this encodes them a lot more implicitly, however the
SKIP_RETURN_CODE feature is the only way to tell ctest that a tests
failure (including a fixtures one) is benign.

Per default ctest will still loudly fail if a test fails for the reason
that a pre-requisite is not there, alerting the user with a printed
warning what causes the test failure.

Specifically for the CI, there is LO2S_TESTS_MAY_SKIP, which is
intentionally undocumented and which makes unexecutable tests skip.

+++ cmake/tests/pfm_counters/pfm_counters.sh
@@ -1,19 +1,42 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
+SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
+
+if ! bash $SCRIPT_DIR/../paranoid.sh 0; then
+	echo "PFM events test needs kernel.perf_event_paranoid=0" >&2
+	exit 127
+fi
+
+if ! command -v perf; then
+	echo "PFM events test requires 'perf' tool!"
+	exit 127
+fi
+
+if ! command -v jq; then
+	echo "PFM events test requires 'jq' tool!"
+	exit 127
+fi
+
+if ! perf list --json 2>&1 >/dev/null; then
+	echo "PFM events test requires 'perf list' with --json support!"
+	exit 127
+fi
+
+pfm_event="$(perf list --json | jq -r 'last(.[] | select(.EventType == "PFM event")) | .EventName')"
+
+if test -z $pfm_event; then
+	echo "perf not build with libpfm support or no PFM event found!"
+	exit 127
+fi
+
 rm -rf test_trace
 
-# select first pfm event (sed -n '3 p')
-# That is not process mode only (grep -v *)
+./lo2s -a --metric-count 1 --metric-leader cpu-cycles -E $pfm_event --output-trace test_trace -- sleep 1
 
-pfm_event=$(./lo2s --list-events | grep -v \* | grep pfm -A 5 | sed -n '3 p')
-
-./lo2s -a --metric-count 1 --metric-leader cpu-cycles -E $pfm_event  --output-trace test_trace -- sleep 1
-
-if  otf2-print test_trace/traces.otf2 | grep  "METRIC" | grep $pfm_event > /dev/null;
-then
-    exit 0
+if otf2-print test_trace/traces.otf2 | grep "METRIC" | grep $pfm_event >/dev/null; then
+	exit 0
 else
-    echo "Trace does not contain counter metrics for ${pfm_event}!"
-    exit 1
+	echo "Trace does not contain counter metrics for ${pfm_event}!"
+	exit 1
 fi
